Imports reorganized to make unit tests independent
