--- app/result.py
@@ -1,7 +1,0 @@
-from hecuba.storageobj import StorageObj
-
-class Result(StorageObj):
-    '''
-    @ClassField instances dict<<word:str>,instances:int>
-    '''
-    pass
