+++ tests/class_definitions.py
@@ -6,3 +6,10 @@ class Words(StorageObj):
     @ClassField words dict<<position:int>,wordinfo:str>
     '''
     pass
+
+
+class TestStorageObj(StorageObj):
+    '''
+       @ClassField test dict<<position:int>,text:str>
+    '''
+    pass
