+++ tests/withcassandra/storage_tests.py
@@ -1,16 +1,14 @@
 import unittest
 
-from app.words import Words
-from hecuba import Config, config
-from storage.api import getByID
-
 
 class StorageTests(unittest.TestCase):
-
     def setUp(self):
+        from hecuba import Config
         Config.reset(mock_cassandra=False)
 
     def test_getByID_block(self):
+        from storage.api import getByID
+        from class_definitions import Words
         # ki = KeyIter('testspace', 'tt', 'app.words.Words', 'fake-id', ['position'])
         SO = Words('so')
         b = SO.split().next()
@@ -19,6 +17,8 @@ class StorageTests(unittest.TestCase):
         self.assertEqual(b, new_block)
 
     def test_getByID_storage_obj(self):
+        from storage.api import getByID
+        from class_definitions import Words
         b = Words('testspace.tt')
         new_block = getByID(b.getID())
         self.assertEqual(b, new_block)
