--- tests/withcassandra/test2storageobj.py
@@ -1,9 +1,0 @@
-from hecuba.storageobj import StorageObj
-
-
-class Test2StorageObj(StorageObj):
-    '''
-       @ClassField name str
-       @ClassField age int
-    '''
-    pass
